+++ src/utils/query-keys.ts
@@ -4,3 +4,4 @@ export const QUERY_KEY_CONDITION_HISTORY = "conditionHistory";
 export const QUERY_KEY_PATIENT = "patient";
 export const QUERY_KEY_PATIENT_MEDICATIONS = "medications";
 export const QUERY_KEY_PATIENT_BUILDER_MEDICATIONS = "builder-medications";
+export const QUERY_KEY_MEDICATION_HISTORY = "medication-history";